requirements.txt
Django==2.1
djangorestframework==3.10.0
django-redis==4.10.0
